Fix mobile pair link back navigation
