+++ mobile/app/_layout.tsx
@@ -10,6 +10,7 @@ import { OrcaLogo } from '../src/components/OrcaLogo'
 import { RpcClientProvider } from '../src/transport/client-context'
 import { getNotificationNavigationPath } from '../src/notifications/notification-routing'
 import { loadHosts } from '../src/transport/host-store'
+import { extractPairingCodeFromUrl } from '../src/transport/pairing'
 
 // Why: keeps the native splash screen visible until the React tree is mounted
 // and ready to render. Without this the user sees a blank white/black frame
@@ -30,25 +31,6 @@ Notifications.setNotificationHandler({
   })
 })
 
-// Why: extract the path+payload that follows the orca://pair anchor so we
-// can route it to the confirm screen. Accept either a hash payload
-// (`orca://pair#<base64>`, the QR / shared form) or a query param
-// (`orca://pair?code=<...>`, future-proof for share sheets that strip
-// fragments).
-function extractPairCode(url: string): string | null {
-  if (!url.startsWith('orca://pair')) return null
-  const hashIndex = url.indexOf('#')
-  if (hashIndex !== -1) {
-    return url.slice(hashIndex + 1) || null
-  }
-  const queryIndex = url.indexOf('?')
-  if (queryIndex !== -1) {
-    const params = new URLSearchParams(url.slice(queryIndex + 1))
-    return params.get('code')
-  }
-  return null
-}
-
 export default function RootLayout() {
   const router = useRouter()
   const handledNotificationIdsRef = useRef<Set<string>>(new Set())
@@ -60,9 +42,12 @@ export default function RootLayout() {
   // covers warm-start (link tapped while app is in memory).
   useEffect(() => {
     function handleUrl(url: string) {
-      const code = extractPairCode(url)
+      const code = extractPairingCodeFromUrl(url)
       if (code) {
-        router.push({ pathname: '/pair-confirm', params: { code } })
+        // Why: Android camera launches can leave Expo Router's unmatched
+        // `orca://pair` route underneath this screen; replacing keeps cancel
+        // and edge-back from revealing the router error page.
+        router.replace({ pathname: '/pair-confirm', params: { code } })
       }
     }
